Fixing issue 549 (#559)
* Fixing issue 549

* Updating readme.

* tests should be actual tests.
